+++ src/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [v20171024]
+
+  - Fix crash on startup, refs #94
+  - Fix crash on null account deletion, refs #90
+  - Fix crash on account deletion twice, refs #88
+  - Fix crash on dialog dismiss, refs #89
+
 ## [v20171022]
 
   - Add flash QR Code support, refs #85